Start of Character Attribute Data
Added a Character script with access to health and attributes. We are
utilizing three attributes for now (Attack, Defense, and Speed).
Scriptable Objects were created to allow us to create a simple object
that is manipulable in Unity and we can assign some default values for
levels and attributes for testing. In the future, this may be taken by a
function.

To access this demo, open the CharacterAttributeTesting scene.